More work on the kernel.  The kernel now has all the basic functionality working.  It does not yet have MOVE or PUSH FORWARD implemented.
The kernel client (RemoteKernel) implements everything that the kernel
does.  I have also added dictionary style access to the kernel's namespace.

The InteractiveCluster class now provide a semi-complete interface for 
working with multiple kernels at the same time.  The basics are there,
but there is little error checking and I have not checked to see if 
all the conventions are consistent.


git-svn-id: http://ipython.scipy.org/svn/ipython@790