I have created a ThreadlessKernelTCPFactory that runs the kernel without threads and without a queue.  This makes it possible to use Twisted in the users namespace in the kernel.
This can be enabled by giving the --threadless flag to the scripts/ipkernel 
script.  Other than that the resulting kernel is the same and can be 
contacted using the existing RemoteKernel and InteractiveCluster classes.  

I have also added a block option to the RemoteKernel.execute() method
that blocks and returns the result of executing the command.



git-svn-id: http://ipython.scipy.org/svn/ipython@1133