Fixed the autopx magic function.
Added a block flag to the execute() methods of RemoteKernel and 
InteractiveCluster.  This flag causes execute() to be run in a blocking
manner and stdin/stdout/stderr to be printed locally.  This makes it
possible to see results w/o starting ipresults.  This is helpful if
you can't receive UDP packets.

The %ps and %autopx magics set block=True by default.


git-svn-id: http://ipython.scipy.org/svn/ipython@1137