More cleanup and bug fixes.
The fooAll methods in RemoteController should not be dynamically generated and I have taken this implementation detail out.  It was causing a bug in scatterAll and the interface to RemoteController is really not the same as that in IMultiEngine.

Added a new tools module.

Fixed some old style names.

Made more classes inherit from object.

git-svn-id: http://ipython.scipy.org/svn/ipython@1796