Moved the multienginepb and vanilla folders to the snapshot_r2111 folder.  These folder contain the multienginepb, enginevanilla and multienginevanilla modules that are outdated as of r2111.  However, these files could still be useful as we do new things.  It is worth keeping a snapshot like this.
As of r2111 the vanilla engine stuff mostly works, but a few tests are skipped.  The multienginevanilla 
hasn't been used for a while and probably doesn't work.

The multienginepb client uses the blockOn module which we don't want to rely on.  It currently doesn't
work, but the version we had a few days ago worked mostly fine. 


git-svn-id: http://ipython.scipy.org/svn/ipython@2114