Merging changes from the sprint1 branch.  The following svn commands were used to do this:
  521  svn merge -r 2286:2287 http://ipython.scipy.org/svn/ipython/ipython/branches/sprint1
  527  svn merge -r 2287:2292 http://ipython.scipy.org/svn/ipython/ipython/branches/sprint1
  528  svn merge -r 2293:2294 http://ipython.scipy.org/svn/ipython/ipython/branches/sprint1
  529  svn merge -r 2294:2296 http://ipython.scipy.org/svn/ipython/ipython/branches/sprint1
  530  svn merge -r 2297:2298 http://ipython.scipy.org/svn/ipython/ipython/branches/sprint1
  531  svn merge -r 2298:2299 http://ipython.scipy.org/svn/ipython/ipython/branches/sprint1

These changes include most of the work done at the IPython sprint (in the sprint1 branch), but not the security related things commited in r2293.  Those things need to be refactored to get tests to pass before merging into saw.

After doing the merge, the IPython1 tests suite would not run.  This was due to a problem in the core1/test_all_doctests.py module.  I commented out the offending things and the tests now pass.  

git-svn-id: http://ipython.scipy.org/svn/ipython@2305